feat: add logic editor components and refactor existing code
This commit introduces new components for the logic editor, including `Shortcut`, `CatchError`, and `ToolBar`, while refactoring existing components like `Button` and `Tran`. The logic editor now supports features such as keyboard shortcuts, error handling, and a toolbar with various actions like zooming, undo/redo, and adding nodes. Additionally, the commit includes style adjustments and improves code maintainability by organizing related files into appropriate directories. The changes enhance the user experience and provide a more robust foundation for future development.
