+++ app/globals.css
@@ -67,7 +67,7 @@
     --accent: 12 6.5% 15.1%;
     --accent-foreground: 60 9.1% 97.8%;
 
-    --border: 240 2% 13%;
+    --border: 240 2% 20%;
     --input: 12 6.5% 15.1%;
     --ring: 24 5.7% 82.9%;
